feat: auth api 정의
