+++ src/api/endpoints/auth.ts
@@ -1,0 +1,5 @@
+export const auth = {
+  login: '/members/login',
+  reissue: '/members/login/refresh',
+  logout: '/members/logout',
+}
