+++ src/api/fetches/index.ts
@@ -4,8 +4,10 @@ import { point } from './point'
 import { quest } from './quest'
 import { streak } from './streak'
 import { survey } from './survey'
+import { auth } from './auth'
 
 export const fetches = {
+  auth,
   member,
   interest,
   point,
