+++ src/api/types/requests/auth.ts
@@ -1,0 +1,6 @@
+interface LoginRequest {
+  email: string
+  password: string
+}
+
+export type { LoginRequest }
