+++ src/api/types/responses/auth.ts
@@ -1,0 +1,9 @@
+interface LoginResponse {
+  accessToken: string
+}
+
+interface ReissueResponse {
+  accessToken: string
+}
+
+export type { LoginResponse, ReissueResponse }
